pyproject.toml
[tool.poetry]
name = "githubmodels"
version = "0.1.0"
description = ""
authors = ["Shuichi Kataoka <shu01k9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
openai = "^1.47.0"
azure-ai-inference = "^1.0.0b4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
